Address trace refactor review feedback
