+++ crates/rulemorph/src/transform/v2_trace.rs
@@ -1,8 +1,5 @@
 use super::*;
-use crate::v2_operator::{
-    operator_has_eager_args, operator_has_item_level_trace, operator_has_lazy_arg_trace,
-    operator_skips_args_when_pipe_is_missing, operator_stops_after_missing_arg,
-};
+use crate::v2_operator::{V2OperatorMetadata, V2OperatorTrace, operator};
 
 pub(super) fn eval_v2_pipe_traced<'a>(
     pipe: &V2Pipe,
@@ -79,8 +76,12 @@ fn eval_v2_step_traced<'a>(
                 .input_v2_eval_value(&pipe_value, collector.options(), None)
                 .attr_count("arg_count", op.args.len())
                 .finish(collector);
-            let result = if operator_has_item_level_trace(&op.op) {
-                eval_v2_collection_op_traced(
+            let operator_metadata = operator(&op.op);
+            let trace = operator_metadata
+                .map(|metadata| metadata.trace)
+                .unwrap_or(V2OperatorTrace::EagerArgs);
+            let result = match trace {
+                V2OperatorTrace::ItemLevelCollection => eval_v2_collection_op_traced(
                     op,
                     pipe_value.clone(),
                     record,
@@ -89,9 +90,19 @@ fn eval_v2_step_traced<'a>(
                     step_path,
                     ctx,
                     collector,
-                )
-            } else if operator_has_eager_args(&op.op) {
-                eval_v2_eager_op_traced(
+                ),
+                V2OperatorTrace::EagerArgs => eval_v2_eager_op_traced(
+                    op,
+                    pipe_value.clone(),
+                    record,
+                    context,
+                    out,
+                    step_path,
+                    ctx,
+                    operator_metadata,
+                    collector,
+                ),
+                V2OperatorTrace::LazyShortCircuit => eval_v2_lazy_op_traced(
                     op,
                     pipe_value.clone(),
                     record,
@@ -100,20 +111,10 @@ fn eval_v2_step_traced<'a>(
                     step_path,
                     ctx,
                     collector,
-                )
-            } else if operator_has_lazy_arg_trace(&op.op) {
-                eval_v2_lazy_op_traced(
-                    op,
-                    pipe_value.clone(),
-                    record,
-                    context,
-                    out,
-                    step_path,
-                    ctx,
-                    collector,
-                )
-            } else {
-                eval_v2_op_step(op, pipe_value.clone(), record, context, out, step_path, ctx)
+                ),
+                V2OperatorTrace::Delegated => {
+                    eval_v2_op_step(op, pipe_value.clone(), record, context, out, step_path, ctx)
+                }
             };
             let output = match result {
                 Ok(output) => output,
@@ -449,16 +450,19 @@ fn eval_v2_eager_op_traced<'a>(
     out: &'a JsonValue,
     step_path: &str,
     ctx: &V2EvalContext<'a>,
+    operator_metadata: Option<&'static V2OperatorMetadata>,
     collector: &mut TraceCollector,
 ) -> Result<V2EvalValue, TransformError> {
     if matches!(pipe_value, V2EvalValue::Missing)
-        && operator_skips_args_when_pipe_is_missing(&op.op)
+        && operator_metadata.is_some_and(|metadata| metadata.skips_args_when_pipe_is_missing)
     {
         return eval_v2_op_step(op, pipe_value, record, context, out, step_path, ctx);
     }
 
     let step_ctx = ctx.clone().with_pipe_value(pipe_value.clone());
     let mut arg_values = Vec::with_capacity(op.args.len());
+    let stops_after_missing_arg =
+        operator_metadata.is_some_and(|metadata| metadata.stops_after_missing_arg);
     for (arg_index, arg) in op.args.iter().enumerate() {
         let arg_path = format!("{}.args[{}]", step_path, arg_index);
         let value =
@@ -466,7 +470,7 @@ fn eval_v2_eager_op_traced<'a>(
         emit_v2_arg_eval(collector, &arg_path, arg_index, &op.op, &value);
         let is_missing = matches!(value, V2EvalValue::Missing);
         arg_values.push(value);
-        if is_missing && operator_stops_after_missing_arg(&op.op) {
+        if is_missing && stops_after_missing_arg {
             break;
         }
     }
